Enhancements to page number text box
o When the user gives focus to the text box, the text in the text box is
selected.
o When the user gives focus to the text box, the text box remembers the
value in the text box. The user then can hit Escape while the text box
has focus to restore the old value.